requirements.txt
html5lib==0.9999999
selenium
django==1.11.*
pymongo>=3.7
py2-py3-django-email-as-username==1.7.1
diff_match_patch
dateutils
bs4
bleach==1.4.2
rauth
elasticsearch
elasticsearch_dsl
regex
pytest
mailchimp
redis
django-redis
django_mobile
roman
p929
git+https://github.com/ByteInternet/python-varnishadm.git@74555895719f7db4c03298a5e54c689db334a903
geojson
django-recaptcha
unicodecsv
pathos
oauth2client
google-api-python-client
jsonpickle
boto3
pillow
datrie
deepdiff==3.3.0
Appium-Python-Client
django-webpack-loader
django-anymail
pyyaml
ua-parser
user-agents
django-user-agents
psycopg2
djangorestframework
djangorestframework_simplejwt==3.3.0
undecorated
apscheduler
pytz
google-cloud-storage
cython
django-easy-timezones
gunicorn
httplib2
lxml
convertdate
ipython
tqdm